30_error_except_json/2_password_manager/requirements.txt
pillow==12.0.0
pyperclip==1.11.0
